==> 87.LimitOfDetection/03.Signal_prediction_and_plot.Rmd <==
---
title: "AANanopore"
author: "Chao Tang"
date: "2022/7/26"
output: html_document
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(eval      = T, # 在块中运行代码(default = TRUE)
                      highlight = T, # 高亮显示
                      echo      = F, # 是否在输出中包含源代码
                      tidy      = T, # 是否整理代码
                      error     = T, # 是否在输出中包含错误信息
                      warning   = F, # 是否在输出中包含警告(default = TRUE)
                      message   = F, # 是否在输出中包含参考的信息
                      cache.    = F)
knitr::opts_knit$set(root.dir = "/mnt/raid61/Personal_data/tangchao/AANanopore")
```

```{r warning=FALSE}
library(data.table)
library(parallel)
library(caret)
library(IRanges)
library(shiny)
library(plotly)
library(ggpubr)
library(ggplot2)
library(patchwork)
```

```{r}
aa_cols <- c(RColorBrewer::brewer.pal(n = 8, "Accent")[-c(1, 4, 5, 8)], 
             RColorBrewer::brewer.pal(n = 11, "PiYG")[c(4, 9, 11)], 
             ggsci::pal_aaas()(10), 
             ggsci::pal_locuszoom()(7)[-c(3, 6, 7)], 
             RColorBrewer::brewer.pal(n = 3, name = "Set2")[1:2])
aa_cols <- plyr::mapvalues(aa_cols, "#D43F3AFF", RColorBrewer::brewer.pal(n = 8, name = "BrBG")[1])
names(aa_cols) <- c(AMINO_ACID_CODE[1:20], "CbC", "His1", "His2")
names(aa_cols) <- c(AMINO_ACID_CODE[1:20], "CMC", "His1", "His2")
```

```{r}
KnnSelecting <- function(signals, dists, dist_method = "euclidean", k = 20, noise = 0) {
  stopifnot(noise < k)
  if(is(dists, "dist")) {
    knn1 <- dbscan::kNN(dists, k = k)
    knn2 <- apply(knn1$id, 2, function(x) x %in% which(attr(dists, "Labels") %in% grep("^Noise", rownames(knn1$id), value = TRUE)))
  } else {
    knn1 <- dbscan::kNN(dists[[dist_method]], k = k)
    knn2 <- apply(knn1$id, 2, function(x) x %in% which(attr(dists[[dist_method]], "Labels") %in% grep("^Noise", rownames(knn1$id), value = TRUE)))
  }
  
  tu <- row.names(knn1$id)[which(rowSums(knn2) <= noise)]
  # tu <- union(tu, knn2[which(rowSums(knn2) <= noise), 1])
  tu <- grep("^Noise", tu, invert = TRUE, value = TRUE)
  ggplot(signals, aes(x = Blockade, y = DwellTime, colour = !ID %in% tu)) + 
    geom_point(size = .5) + 
    scale_y_log10() + 
    scale_colour_manual(values = c("#E41A1C", "#377EB8")) + 
    geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[c("E", "F", "L")], Blockade]) + 
    theme_classic() + 
    theme(legend.position = "none") -> p1
  ggplot(signals[ID %in% tu], aes(x = Blockade, y = DwellTime, colour = ID %in% tu)) + 
    geom_point(size = .5) + 
    scale_y_log10() + 
    scale_colour_manual(values = "#E41A1C") + 
    geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[c("E", "F", "L")], Blockade]) + 
    theme_classic() + 
    theme(legend.position = "none") + 
    labs(title = length(tu)) -> p2
  ggplot(signals[!ID %in% tu], aes(x = Blockade, y = DwellTime, colour = ID %in% tu)) + 
    geom_point(size = .5) + 
    scale_y_log10() + 
    scale_colour_manual(values = "#377EB8") + 
    geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[c("E", "F", "L")], Blockade]) + 
    theme_classic() + 
    theme(legend.position = "none") -> p3
  print(p1 + p2 + p3)
  return(tu)
}

RemoveOutlier <- function(tree, n = 10, Keep = 0.95) {
  h <- length(tree$height)
  g <- cutree(tree, h = tree$height[h])
  g <- data.table(tip = names(g), cluster = g)
  
  while (g[, .N, cluster][N >= n, sum(N)/length(tree$labels)] > Keep) {
    h <- h - 1
    g <- cutree(tree, h = tree$height[h])
    g <- data.table(tip = names(g), cluster = g)
  }
  cat(paste0("Keep: ", round(g[, .N, cluster][N >= n, sum(N)/length(tree$labels)] * 100, 2), "%"))
  g[cluster %in% g[, .N, cluster][N >= n, cluster], tip]
}

get_density <- function(x, y, n = 1000, ...) {
  dens <- MASS::kde2d(x, y, n = n, ...)
  ix <- findInterval(x, dens$x)
  iy <- findInterval(y, dens$y)
  ii <- cbind(ix, iy)
  return(dens$z[ii] / max(dens$z[ii]))
}
```

```{r}
meta0 <- as.data.table(openxlsx::read.xlsx("data/MetaInfomation/整理数据(20230831).xlsx", sheet = 3))
colnames(meta0) <- c("file_name", "start_time", "end_time", "amino_acid", "file_id", "purpose", "sample", "base_line", "note")
meta <- unique(meta0[, 1:4])
setkey(meta, file_name, start_time)
meta <- meta[, .SD[, .(amino_acid, start_time, end_time, file_id = paste0(file_name, ".", seq_len(.N)))], file_name]
meta <- meta[file_id %in% gsub(".MainL0.txt", "", list.files("./analysis/87.LimitOfDetection/01.SelectedL0"))]
meta <- meta[grepl("2023092", file_name)]
```

```{r}
meta[, sig_file := paste0("./analysis/87.LimitOfDetection/01.SelectedL0/", file_id, ".MainL0.txt")]
```

```{r}
sigs <- do.call(rbind, lapply(meta$sig_file, fread))
```

```{r fig.width=10, fig.height=12}
ggplot(sigs, aes(x = Blockade, y = DwellTime, colour = A)) + 
  geom_point() + 
  scale_y_log10() + 
  facet_wrap(~ A, ncol = 1, scales = "free_y")
```

## Distance of signal

### Background data

```{r}
bgsignals0 <- do.call(rbind, lapply(list.files("./analysis/81.ABFProcessing/SelectedSignals", "_background.txt", full.names = TRUE), fread))
bgsignals0[, ID := paste0("Noise_", ID)]
# bgsignals0 <- bgsignals0[DwellTime > 0.3 & Blockade < 0.3]

Big_FeatureMatrixs <- readRDS("./analysis/81.ABFProcessing/SelectedSignals/01.StandardAA/01.SignalDistance/StandardAA_Big_FeatureMatrixs.Rds")
Big_FeatureMatrixs <- Big_FeatureMatrixs[ID %in% bgsignals0$ID]
Big_FeatureMatrixs[, AA := NULL]
```

```{r}
bgsignals <- sigs[A %in% meta[is.na(amino_acid), file_id], ]
bgsignals[, ID := paste0("Noise_", ID)]
# bgsignals <- bgsignals[DwellTime > 0.3 & Blockade < 0.3]

B_file <- bgsignals[, as.character(unique(A))]
Big_FeatureMatrix <- lapply(B_file, function(x) {
  B_file <- unlist(strsplit(x, "\\."))[1]
  Big_FeatureMatrix <- list.files(path = "./analysis/81.ABFProcessing/FeatureMatrix", B_file, full.names = TRUE)
  Big_FeatureMatrix <- fread(Big_FeatureMatrix)
  Big_FeatureMatrix[, ID := paste0("Noise_", ID)]
  Big_FeatureMatrix[ID %in% bgsignals$ID]
})
Big_FeatureMatrix <- do.call(rbind, Big_FeatureMatrix[1:2])
```

```{r}
Big_FeatureMatrixs <- Big_FeatureMatrixs[!ID %in% Big_FeatureMatrix[, ID]]
```

### Signal data

```{r}
sgsignals <- sigs[A %in% meta[!is.na(amino_acid), file_id], ]
# sgsignals <- sgsignals[DwellTime > 0.3 & Blockade < 0.3]

B_file <- sgsignals[, as.character(unique(A))]
B_file <- unique(mapply(function(x) unlist(strsplit(x, "\\."))[1], B_file))
Sig_FeatureMatrix <- lapply(B_file, function(x) {
  # print(x)
  fread(list.files(path = "./analysis/81.ABFProcessing/FeatureMatrix", x, full.names = TRUE))[ID %in% sgsignals$ID]
})
Sig_FeatureMatrix <- do.call(rbind, Sig_FeatureMatrix)
```

```{r eval=FALSE}
mclapply(sgsignals[, as.character(unique(A))], function(x) {
  # if(file.exists(paste0("./analysis/87.LimitOfDetection/02.SignalsDistance/", x, "_Signals_euclidean_distance.Rds"))) return(NULL)
  SFMi <- Sig_FeatureMatrix[ID %in% sgsignals[A == x, ID]]
  if(nrow(SFMi) > nrow(Big_FeatureMatrix)) {
    FeatureMatrix <- rbind(Big_FeatureMatrix, 
                           Big_FeatureMatrixs[sample(.N, nrow(SFMi) - nrow(Big_FeatureMatrix))], 
                           SFMi)
  } else {
    FeatureMatrix <- rbind(Big_FeatureMatrix[sample(.N, nrow(SFMi))], SFMi)
  }
  setkey(FeatureMatrix, ID)
  FeatureMatrix <- data.frame(FeatureMatrix[, grepl("^X", colnames(FeatureMatrix)) | grepl("DwellTime", colnames(FeatureMatrix)), with = F], row.names = FeatureMatrix[[1]])
  euclidean = stats::dist(FeatureMatrix, method = "euclidean")
  saveRDS(euclidean, file = paste0("./analysis/87.LimitOfDetection/02.SignalsDistance/", x, "_Signals_euclidean_distance.Rds"))
}, mc.cores = 10)
```

## Distance filtering

```{r}
model <- readRDS("./analysis/81.ABFProcessing/SelectedSignals/01.StandardAA/06.Modeling/Version2/L1/SubAAs/ModelEFL/RFmodel.Rds")
```

```{r}
meta1 <- meta[file_name %in% c("20230920_0001", "20230920_0002", "20230920_0003", "20230920_0004", "20230920_0006", "20230920_0008", "20230920_0009")]
```

```{r}
subtab <- Big_FeatureMatrix[ID %in% bgsignals[A %in% meta1[, file_id], ID]]
pred0 <- data.table(ID = subtab$ID, Pred = predict(model, subtab), Prob = apply(predict(model, subtab, type = "prob"), 1, max), 
                    Delta = apply(predict(model, subtab, type = "prob"), 1, function(x) -diff(head(sort(x, decreasing = T), 2))))
```

```{r}
subtab <- Sig_FeatureMatrix[ID %in% sgsignals[A %in% meta1[, file_id], ID]]
pred1 <- data.table(ID = subtab$ID, Pred = predict(model, subtab), Prob = apply(predict(model, subtab, type = "prob"), 1, max), 
                    Delta = apply(predict(model, subtab, type = "prob"), 1, function(x) -diff(head(sort(x, decreasing = T), 2))))
```

```{r}
res <- rbind(merge(bgsignals[A %in% meta1[, file_id], ], pred0, by = "ID"), merge(sgsignals[A %in% meta1[, file_id], ], pred1, by = "ID"))
res[A == "20230920_0001.1", Concentration := "Blank"]
res[A == "20230920_0002.1", Concentration := "100 nM"]
res[A == "20230920_0003.1", Concentration := "500 nM"]
res[A == "20230920_0004.1", Concentration := "1 μM"]
res[A == "20230920_0006.1", Concentration := "2 μM"]
res[A == "20230920_0008.1", Concentration := "10 μM"]
res[A == "20230920_0009.1", Concentration := "10 μM"]
res[, Concentration := factor(Concentration, levels = c("Blank", "100 nM", "500 nM", "1 μM", "2 μM", "10 μM"))]
```

```{r}
res_list <- split(res, res$Concentration)
res_list <- lapply(res_list, function(x) {
  data.table(x, D = x[, get_density(x = Blockade, y = log10(DwellTime))])
})
res <- do.call(rbind, res_list)
```

```{r}
aa <- do.call(rbind, lapply(paste0("./analysis/81.ABFProcessing/SelectedSignals/01.StandardAA/03.GroupedSignals/", c("Glu", "Phe", "Leu"), ".signal.txt"), fread))[State == "Sington"]
aa <- aa[, mean_sd(Blockade), AA]
```

```{r fig.width=12, fig.height=6}
ggplot() + 
  geom_point(data = res, aes(x = Blockade, y = DwellTime, colour = pmin(D + .2, 1)), size = 0.4) + 
  scale_y_log10(limits = c(0.1, 200)) + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_gradient(high = "black", low = "grey") + 
  geom_rect(data = aa[AA == "Glu"], aes(xmin = ymin, xmax = ymax, ymin = 0, ymax = 100), fill = aa_cols["Glu"], alpha = 0.3) + 
  geom_rect(data = aa[AA == "Phe"], aes(xmin = ymin, xmax = ymax, ymin = 0, ymax = 100), fill = aa_cols["Phe"], alpha = 0.3) + 
  geom_rect(data = aa[AA == "Leu"], aes(xmin = ymin, xmax = ymax, ymin = 0, ymax = 100), fill = aa_cols["Leu"], alpha = 0.3) + 
  geom_text(data = aa, aes(x = y, y = 180, label = AA)) + 
  labs(tag = "a") + 
  theme_bw(base_size = 15) + 
  theme(legend.position = "none", panel.grid = element_blank(), plot.tag = element_text(size = 20, face = "bold")) + 
  facet_wrap(~ Concentration, nrow = 2) -> p1
p1
```

```{r fig.width=12, fig.height=3}
ggplot() + 
  geom_point(data = res[Concentration != "Blank"][Delta > 0.99 & Prob > 0.99], aes(x = Blockade, y = DwellTime, colour = Pred)) + 
  scale_y_log10(limits = c(0.1, 200)) + 
  geom_rect(data = aa[AA == "Glu"], aes(xmin = ymin, xmax = ymax, ymin = 0, ymax = 100), fill = aa_cols["Glu"], alpha = 0.3) + 
  geom_rect(data = aa[AA == "Phe"], aes(xmin = ymin, xmax = ymax, ymin = 0, ymax = 100), fill = aa_cols["Phe"], alpha = 0.3) + 
  geom_rect(data = aa[AA == "Leu"], aes(xmin = ymin, xmax = ymax, ymin = 0, ymax = 100), fill = aa_cols["Leu"], alpha = 0.3) + 
  geom_text(data = aa, aes(x = y, y = 180, label = AA)) + 
  scale_colour_manual(breaks = names(aa_cols), values = aa_cols) + 
  labs(tag = "b") + 
  theme_bw(base_size = 15) + 
  theme(legend.position = "none", panel.grid = element_blank(), plot.tag = element_text(size = 20, face = "bold")) + 
  facet_wrap(~ Concentration, nrow = 1) -> p2
p2
```

```{r fig.width=10, fig.height=9}
cowplot::plot_grid(p1, p2, ncol = 1, rel_heights = c(2, 1))
```

```{r warning=FALSE}
cairo_pdf("./analysis/00.FigureTables/Rebuttal/NatureMethods/LOD.pdf", width = 10, height = 9)
cowplot::plot_grid(p1, p2, ncol = 1, rel_heights = c(2, 1))
dev.off()
```
