==> 87.LimitOfDetection/13.SelectedSignals_Asp.Rmd <==
---
title: "Signal selecting"
author: "Chao Tang"
date: 'Report created: `r Sys.Date()`'
output: 
  html_document: 
    code_folding: "hide"
    toc: true
    toc_depth: 4
    toc_float: 
      collapsed: false
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, 
                      warning = FALSE, 
                      message = FALSE, 
                      fig.align = 'center')
knitr::opts_knit$set(root.dir = "/mnt/raid61/Personal_data/tangchao/AANanopore")
```

```{r}
library(factoextra)
library(data.table)
library(Biostrings)
library(patchwork)
library(ggplot2)
library(ggrepel)
library(IRanges)
library(dplyr)
library(lsa)
```

```{r}
KnnSelecting <- function(signals, dists, dist_method = "euclidean", k = 20, noise = 0, index = NULL) {
  stopifnot(noise < k)
  if(is(dists, "dist")) {
    knn1 <- dbscan::kNN(dists, k = k)
    knn2 <- apply(knn1$id, 2, function(x) x %in% which(attr(dists, "Labels") %in% grep("^Noise", rownames(knn1$id), value = TRUE)))
  } else {
    knn1 <- dbscan::kNN(dists[[dist_method]], k = k)
    knn2 <- apply(knn1$id, 2, function(x) x %in% which(attr(dists[[dist_method]], "Labels") %in% grep("^Noise", rownames(knn1$id), value = TRUE)))
  }
  
  tu <- row.names(knn1$id)[which(rowSums(knn2) <= noise)]
  # tu <- union(tu, knn2[which(rowSums(knn2) <= noise), 1])
  tu <- grep("^Noise", tu, invert = TRUE, value = TRUE)
  ggplot(signals, aes(x = Blockade, y = DwellTime, colour = !ID %in% tu)) + 
    geom_point(size = .5) + 
    scale_y_log10() + 
    scale_colour_manual(values = c("#E41A1C", "#377EB8")) + 
    theme_classic() + 
    theme(legend.position = "none") -> p1
  ggplot(signals[ID %in% tu], aes(x = Blockade, y = DwellTime, colour = ID %in% tu)) + 
    geom_point(size = .5) + 
    scale_y_log10() + 
    scale_colour_manual(values = "#E41A1C") + 
    theme_classic() + 
    theme(legend.position = "none") + 
    labs(title = length(tu)) -> p2
  ggplot(signals[!ID %in% tu], aes(x = Blockade, y = DwellTime, colour = ID %in% tu)) + 
    geom_point(size = .5) + 
    scale_y_log10() + 
    scale_colour_manual(values = "#377EB8") + 
    theme_classic() + 
    theme(legend.position = "none") -> p3
  if(!is.null(index)) {
    p1 <- p1 + geom_vline(xintercept = index, colour = "red")
    p2 <- p2 + geom_vline(xintercept = index, colour = "red")
    p3 <- p3 + geom_vline(xintercept = index, colour = "red")
  }
  print(p1 + p2 + p3)
  return(tu)
}

RemoveOutlier <- function(tree, n = 10, Keep = 0.95) {
  h <- length(tree$height)
  g <- cutree(tree, h = tree$height[h])
  g <- data.table(tip = names(g), cluster = g)
  
  while (g[, .N, cluster][N >= n, sum(N)/length(tree$labels)] > Keep) {
    h <- h - 1
    g <- cutree(tree, h = tree$height[h])
    g <- data.table(tip = names(g), cluster = g)
  }
  cat(paste0("Keep: ", round(g[, .N, cluster][N >= n, sum(N)/length(tree$labels)] * 100, 2), "%"))
  g[cluster %in% g[, .N, cluster][N >= n, cluster], tip]
}
```

```{r}
AABlockade <- lapply(list.files("./analysis/61.SignalSelecting/01.StandardAA/02.SelectedSignals", "_V2.Rds", full.names = T), function(x) readRDS(x)$Summary)
AABlockade <- do.call(rbind, AABlockade)
AABlockade <- AABlockade[State == "State1"]
```

```{r}
meta <- as.data.table(openxlsx::read.xlsx("data/MetaInfomation/氨基酸浓度梯度实验信息表.xlsx", sheet = 1))
colnames(meta) <- c("file_name", "concentration", "amino_acid")
meta[, file_name := gsub(".abf", "", file_name)]
setkey(meta, file_name)
meta <- meta[, .SD[, .(amino_acid, concentration, file_id = paste0(file_name, ".", seq_len(.N)))], file_name]
meta <- meta[file_id %in% gsub(".MainL0.txt", "", list.files("./analysis/87.LimitOfDetection/01.SelectedL0"))]
meta[, sig_file := paste0("./analysis/87.LimitOfDetection/01.SelectedL0/", file_id, ".MainL0.txt")]
meta[, file.exists(sig_file)]
meta$amino_acid <- plyr::mapvalues(meta$amino_acid, tolower(AMINO_ACID_CODE), names(AMINO_ACID_CODE))
meta <- meta[amino_acid == "D"]
meta[concentration == "0", amino_acid := NA]
meta
```

# D

```{r}
meta[amino_acid == "D"]
```

```{r fig.width=24, fig.height=6}
dataset <- do.call(rbind, lapply(meta[amino_acid == "D", sig_file], fread))[Blockade < 0.35 & DwellTime > 0.3]
B_file <- paste0("./analysis/81.ABFProcessing/FeatureMatrix/FeatureMatrix_", unique(meta[amino_acid == "D", file_name]), ".txt")
FeatureMatrix <- do.call(rbind, lapply(B_file, fread))
FeatureMatrix <- FeatureMatrix[ID %in% dataset$ID]
FeatureMatrix <- data.frame(FeatureMatrix[, grepl("^X", colnames(FeatureMatrix)) | grepl("DwellTime", colnames(FeatureMatrix)), with = F], row.names = FeatureMatrix[[1]])
```

## 21121003.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/87.LimitOfDetection/02.SignalsDistance/21121003.1.signal.dist.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "21121003.1"], dists = alldists, k = 20, noise = 0, index = AABlockade[AA == "Asp", Blockade])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(subtab)
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 10, Keep = 0.9)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r fig.width=6, fig.height=4}
ggplot(dataset[A == "21121003.1"], aes(x = Blockade, y = DwellTime, colour = !ID %in% sigs)) + 
  geom_point(size = 0.4) + 
  scale_y_log10() + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("D", ""))], Blockade], col = 2) + 
  theme_base() + theme(legend.position = "none", plot.background = element_blank())
```

```{r}
fwrite(dataset[ID %in% sigs], file.path("./analysis/87.LimitOfDetection/03.SelectedSignals/D.21121003.1.selected.signals.txt"), sep = "\t", quote = F)
```

## 21121005.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/87.LimitOfDetection/02.SignalsDistance/21121005.1.signal.dist.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "21121005.1"], dists = alldists, k = 20, noise = 0, index = AABlockade[AA == "Asp", Blockade])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(subtab)
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 10, Keep = 0.99)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r fig.width=6, fig.height=4}
ggplot(dataset[A == "21121005.1"], aes(x = Blockade, y = DwellTime, colour = !ID %in% sigs)) + 
  geom_point(size = 0.4) + 
  scale_y_log10() + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("D", ""))], Blockade], col = 2) + 
  theme_base() + theme(legend.position = "none", plot.background = element_blank())
```

```{r}
fwrite(dataset[ID %in% sigs], file.path("./analysis/87.LimitOfDetection/03.SelectedSignals/D.21121005.1.selected.signals.txt"), sep = "\t", quote = F)
```

## 21121006.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/87.LimitOfDetection/02.SignalsDistance/21121006.1.signal.dist.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "21121006.1"], dists = alldists, k = 10, noise = 0, index = AABlockade[AA == "Asp", Blockade])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(subtab)
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 10, Keep = 0.95)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r fig.width=6, fig.height=4}
ggplot(dataset[A == "21121006.1"], aes(x = Blockade, y = DwellTime, colour = !ID %in% sigs)) + 
  geom_point(size = 0.4) + 
  scale_y_log10() + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("D", ""))], Blockade], col = 2) + 
  theme_base() + theme(legend.position = "none", plot.background = element_blank())
```

```{r}
fwrite(dataset[ID %in% sigs & Blockade > 0.2], file.path("./analysis/87.LimitOfDetection/03.SelectedSignals/D.21121006.1.selected.signals.txt"), sep = "\t", quote = F)
```

## 21121007.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/87.LimitOfDetection/02.SignalsDistance/21121007.1.signal.dist.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "21121007.1"], dists = alldists, k = 10, noise = 0, index = AABlockade[AA == "Asp", Blockade])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(subtab)
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 10, Keep = 0.95)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r fig.width=6, fig.height=4}
ggplot(dataset[A == "21121007.1"], aes(x = Blockade, y = DwellTime, colour = !ID %in% sigs)) + 
  geom_point(size = 0.4) + 
  scale_y_log10() + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("D", ""))], Blockade], col = 2) + 
  theme_base() + theme(legend.position = "none", plot.background = element_blank())
```

```{r}
fwrite(dataset[ID %in% sigs & Blockade > 0.2], file.path("./analysis/87.LimitOfDetection/03.SelectedSignals/D.21121007.1.selected.signals.txt"), sep = "\t", quote = F)
```

## 21124008.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/87.LimitOfDetection/02.SignalsDistance/21124008.1.signal.dist.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "21124008.1"], dists = alldists, k = 10, noise = 0, index = AABlockade[AA == "Asp", Blockade])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(subtab)
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 8, Keep = 0.95)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r fig.width=6, fig.height=4}
ggplot(dataset[A == "21124008.1"], aes(x = Blockade, y = DwellTime, colour = !ID %in% sigs)) + 
  geom_point(size = 0.4) + 
  scale_y_log10() + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("D", ""))], Blockade], col = 2) + 
  theme_base() + theme(legend.position = "none", plot.background = element_blank())
```

```{r}
fwrite(dataset[ID %in% sigs & Blockade > 0.2], file.path("./analysis/87.LimitOfDetection/03.SelectedSignals/D.21124008.1.selected.signals.txt"), sep = "\t", quote = F)
```

## 21124009.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/87.LimitOfDetection/02.SignalsDistance/21124009.1.signal.dist.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "21124009.1"], dists = alldists, k = 10, noise = 0, index = AABlockade[AA == "Asp", Blockade])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(subtab)
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 8, Keep = 0.95)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r fig.width=6, fig.height=4}
ggplot(dataset[A == "21124009.1"], aes(x = Blockade, y = DwellTime, colour = !ID %in% sigs)) + 
  geom_point(size = 0.4) + 
  scale_y_log10() + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("D", ""))], Blockade], col = 2) + 
  theme_base() + theme(legend.position = "none", plot.background = element_blank())
```

```{r}
fwrite(dataset[ID %in% sigs & Blockade > 0.2], file.path("./analysis/87.LimitOfDetection/03.SelectedSignals/D.21124009.1.selected.signals.txt"), sep = "\t", quote = F)
```

## 21124010.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/87.LimitOfDetection/02.SignalsDistance/21124010.1.signal.dist.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "21124010.1"], dists = alldists, k = 10, noise = 0, index = AABlockade[AA == "Asp", Blockade])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(subtab)
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 8, Keep = 0.95)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r fig.width=6, fig.height=4}
ggplot(dataset[A == "21124010.1"], aes(x = Blockade, y = DwellTime, colour = !ID %in% sigs)) + 
  geom_point(size = 0.4) + 
  scale_y_log10() + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("D", ""))], Blockade], col = 2) + 
  theme_base() + theme(legend.position = "none", plot.background = element_blank())
```

```{r}
fwrite(dataset[ID %in% sigs & Blockade > 0.2], file.path("./analysis/87.LimitOfDetection/03.SelectedSignals/D.21124010.1.selected.signals.txt"), sep = "\t", quote = F)
```

## 21124011.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/87.LimitOfDetection/02.SignalsDistance/21124011.1.signal.dist.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "21124011.1"], dists = alldists, k = 10, noise = 0, index = AABlockade[AA == "Asp", Blockade])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(subtab)
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 8, Keep = 0.95)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r fig.width=6, fig.height=4}
ggplot(dataset[A == "21124011.1"], aes(x = Blockade, y = DwellTime, colour = !ID %in% sigs)) + 
  geom_point(size = 0.4) + 
  scale_y_log10() + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("D", ""))], Blockade], col = 2) + 
  theme_base() + theme(legend.position = "none", plot.background = element_blank())
```

```{r}
fwrite(dataset[ID %in% sigs & Blockade > 0.2], file.path("./analysis/87.LimitOfDetection/03.SelectedSignals/D.21124011.1.selected.signals.txt"), sep = "\t", quote = F)
```

## 21124024.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/87.LimitOfDetection/02.SignalsDistance/21124024.1.signal.dist.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "21124024.1"], dists = alldists, k = 3, noise = 0, index = AABlockade[AA == "Asp", Blockade])
```

```{r}
sigs <- intersect(sigs, dataset[Blockade > AABlockade[AA == "Asp", Blockade], ID])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(subtab)
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 3, Keep = 1)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r fig.width=6, fig.height=4}
ggplot(dataset[A == "21124024.1"], aes(x = Blockade, y = DwellTime, colour = !ID %in% sigs)) + 
  geom_point(size = 0.4) + 
  scale_y_log10() + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("D", ""))], Blockade], col = 2) + 
  theme_base() + theme(legend.position = "none", plot.background = element_blank())
```

```{r}
fwrite(dataset[ID %in% sigs & Blockade > 0.2], file.path("./analysis/87.LimitOfDetection/03.SelectedSignals/D.21124024.1.selected.signals.txt"), sep = "\t", quote = F)
```

## 21124027.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/87.LimitOfDetection/02.SignalsDistance/21124027.1.signal.dist.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "21124027.1"], dists = alldists, k = 20, noise = 0, index = AABlockade[AA == "Asp", Blockade])
```

```{r}
sigs <- intersect(sigs, dataset[Blockade > 0.18, ID])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(subtab)
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 3, Keep = 0.95)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r fig.width=6, fig.height=4}
ggplot(dataset[A == "21124027.1"], aes(x = Blockade, y = DwellTime, colour = !ID %in% sigs)) + 
  geom_point(size = 0.4) + 
  scale_y_log10() + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("D", ""))], Blockade], col = 2) + 
  theme_base() + theme(legend.position = "none", plot.background = element_blank())
```

```{r}
fwrite(dataset[ID %in% sigs & Blockade > 0.2], file.path("./analysis/87.LimitOfDetection/03.SelectedSignals/D.21124027.1.selected.signals.txt"), sep = "\t", quote = F)
```

## 21124028.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/87.LimitOfDetection/02.SignalsDistance/21124028.1.signal.dist.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "21124028.1"], dists = alldists, k = 30, noise = 0, index = AABlockade[AA == "Asp", Blockade])
```

```{r}
sigs <- intersect(sigs, dataset[Blockade > 0.18, ID])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(subtab)
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 3, Keep = 0.95)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r fig.width=6, fig.height=4}
ggplot(dataset[A == "21124028.1"], aes(x = Blockade, y = DwellTime, colour = !ID %in% sigs)) + 
  geom_point(size = 0.4) + 
  scale_y_log10() + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("D", ""))], Blockade], col = 2) + 
  theme_base() + theme(legend.position = "none", plot.background = element_blank())
```

```{r}
fwrite(dataset[ID %in% sigs & Blockade > 0.2], file.path("./analysis/87.LimitOfDetection/03.SelectedSignals/D.21124028.1.selected.signals.txt"), sep = "\t", quote = F)
```

## 21125013.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/87.LimitOfDetection/02.SignalsDistance/21125013.1.signal.dist.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "21125013.1"], dists = alldists, k = 5, noise = 0, index = AABlockade[AA == "Asp", Blockade])
```

```{r}
sigs <- intersect(sigs, dataset[Blockade > 0.2, ID])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(subtab)
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 3, Keep = 0.95)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r fig.width=6, fig.height=4}
ggplot(dataset[A == "21125013.1"], aes(x = Blockade, y = DwellTime, colour = !ID %in% sigs)) + 
  geom_point(size = 0.4) + 
  scale_y_log10() + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("D", ""))], Blockade], col = 2) + 
  theme_base() + theme(legend.position = "none", plot.background = element_blank())
```

```{r}
fwrite(dataset[ID %in% sigs & Blockade > 0.2], file.path("./analysis/87.LimitOfDetection/03.SelectedSignals/D.21125013.1.selected.signals.txt"), sep = "\t", quote = F)
```

## 21125014.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/87.LimitOfDetection/02.SignalsDistance/21125014.1.signal.dist.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "21125014.1"], dists = alldists, k = 4, noise = 0, index = AABlockade[AA == "Asp", Blockade])
```

```{r}
sigs <- intersect(sigs, dataset[Blockade > 0.2, ID])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(subtab)
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 10, Keep = 0.85)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r fig.width=6, fig.height=4}
ggplot(dataset[A == "21125014.1"], aes(x = Blockade, y = DwellTime, colour = !ID %in% sigs)) + 
  geom_point(size = 0.4) + 
  scale_y_log10() + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("D", ""))], Blockade], col = 2) + 
  theme_base() + theme(legend.position = "none", plot.background = element_blank())
```

```{r}
fwrite(dataset[ID %in% sigs & Blockade > 0.2], file.path("./analysis/87.LimitOfDetection/03.SelectedSignals/D.21125014.1.selected.signals.txt"), sep = "\t", quote = F)
```

## 21125015.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/87.LimitOfDetection/02.SignalsDistance/21125015.1.signal.dist.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "21125015.1"], dists = alldists, k = 4, noise = 0, index = AABlockade[AA == "Asp", Blockade])
```

```{r}
sigs <- intersect(sigs, dataset[Blockade > 0.2, ID])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(subtab)
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 10, Keep = 0.95)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r fig.width=6, fig.height=4}
ggplot(dataset[A == "21125015.1"], aes(x = Blockade, y = DwellTime, colour = !ID %in% sigs)) + 
  geom_point(size = 0.4) + 
  scale_y_log10() + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("D", ""))], Blockade], col = 2) + 
  theme_base() + theme(legend.position = "none", plot.background = element_blank())
```

```{r}
fwrite(dataset[ID %in% sigs & Blockade > 0.2], file.path("./analysis/87.LimitOfDetection/03.SelectedSignals/D.21125015.1.selected.signals.txt"), sep = "\t", quote = F)
```

## 21125017.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/87.LimitOfDetection/02.SignalsDistance/21125017.1.signal.dist.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "21125017.1"], dists = alldists, k = 5, noise = 0, index = AABlockade[AA == "Asp", Blockade])
```

```{r}
sigs <- intersect(sigs, dataset[Blockade > 0.2, ID])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(subtab)
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 10, Keep = 0.97)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r fig.width=6, fig.height=4}
ggplot(dataset[A == "21125017.1"], aes(x = Blockade, y = DwellTime, colour = !ID %in% sigs)) + 
  geom_point(size = 0.4) + 
  scale_y_log10() + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("D", ""))], Blockade], col = 2) + 
  theme_base() + theme(legend.position = "none", plot.background = element_blank())
```

```{r}
fwrite(dataset[ID %in% sigs & Blockade > 0.2], file.path("./analysis/87.LimitOfDetection/03.SelectedSignals/D.21125017.1.selected.signals.txt"), sep = "\t", quote = F)
```

## 21125018.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/87.LimitOfDetection/02.SignalsDistance/21125018.1.signal.dist.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "21125018.1"], dists = alldists, k = 5, noise = 0, index = AABlockade[AA == "Asp", Blockade])
```

```{r}
sigs <- intersect(sigs, dataset[Blockade > 0.2, ID])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(subtab)
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 10, Keep = 0.97)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r fig.width=6, fig.height=4}
ggplot(dataset[A == "21125018.1"], aes(x = Blockade, y = DwellTime, colour = !ID %in% sigs)) + 
  geom_point(size = 0.4) + 
  scale_y_log10() + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("D", ""))], Blockade], col = 2) + 
  theme_base() + theme(legend.position = "none", plot.background = element_blank())
```

```{r}
fwrite(dataset[ID %in% sigs & Blockade > 0.2], file.path("./analysis/87.LimitOfDetection/03.SelectedSignals/D.21125018.1.selected.signals.txt"), sep = "\t", quote = F)
```


# All


```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/87.LimitOfDetection/02.SignalsDistance/D_all_sigs_distance.Rds"))
sigs <- KnnSelecting(signals = dataset, dists = alldists, k = 7, noise = 0, index = AABlockade[AA == "Asp", Blockade])
```

```{r}
sigs <- intersect(sigs, dataset[Blockade > 0.2, ID])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(subtab)
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 10, Keep = 0.9)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r fig.width=6, fig.height=4}
ggplot(dataset, aes(x = Blockade, y = DwellTime, colour = !ID %in% sigs)) + 
  geom_point(size = 0.4) + 
  scale_y_log10() + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("R", ""))], Blockade], col = 2) + 
  theme_base() + theme(legend.position = "none", plot.background = element_blank())
```

```{r}
fwrite(dataset[ID %in% sigs], file.path("./analysis/87.LimitOfDetection/03.SelectedSignals/D.selected.signals.txt"), sep = "\t", quote = F)
```
